% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_cells.R
\docType{data}
\name{alpha_cells}
\alias{alpha_cells}
\title{Healthy Pancreatic Alpha Cells scRNA-seq data}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-5061/}
}
\description{
A dataset containing the gene expression counts for 443 healthy pancreatic
alpha cells. Highly correlated genes have been removed to reduce data size.
}
