% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_estimate_parameters.R
\name{approximate_gene_mean}
\alias{approximate_gene_mean}
\title{Approximate Gene Mean Distribution}
\usage{
approximate_gene_mean(data_summaries)
}
\arguments{
\item{data_summaries}{an R object that has been output by the package's
compute_data_summaries function.}
}
\value{
A vector of length two, where the first number is the shape of the
 gamma distribution and the second number is the rate of the gamma
 distribution
}
\description{
This function estimates the parameters for the gamma distribution
 of gene means.
}
\details{
Prior to estimating the simulation parameters, it is important to run
 the compute_data_summaries function to build an object that is in the right
 format for the following estimation functions to properly compute.The
 'fitdistrplus' package is required for this function to work. Please install
 it
}
\examples{
data_summaries <- compute_data_summaries(clean_expr_data)
gene_mean_params <- approximate_gene_mean(data_summaries)

}
