% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_estimate_parameters.R
\name{approximate_library_sizes}
\alias{approximate_library_sizes}
\title{Approximate Library Size Distribution}
\usage{
approximate_library_sizes(data_summaries)
}
\arguments{
\item{data_summaries}{an R object that has been output by the package's
compute_data_summaries function.}
}
\value{
A vector of length two, where the first number is the mean of the
 lognormal distribution and the second number is the standard deviation of
 the lognormal distribution
}
\description{
This function estimates the parameters for the lognormal
 distribution of the library sizes.
}
\details{
Prior to estimating the simulation parameters, it is important to run
 the compute_data_summaries function to build an object that is in the right
 format for the following estimation functions to properly compute.The
 'fitdistrplus' package is required for this function to work. Please install
 it.
}
\examples{
data_summaries <- compute_data_summaries(clean_expr_data)
library_params <- approximate_library_sizes(data_summaries)

}
