% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mcc.R
\name{calc_mcc}
\alias{calc_mcc}
\title{Efficient function to calculate Matthew's Correlation Co-efficient (MCC)}
\usage{
calc_mcc(predicted, actual, cutoff)
}
\arguments{
\item{predicted}{predicted continuous values as vector}

\item{actual}{actual values (0,1)}

\item{cutoff}{cut-off for calculation, made for p-values so backwards to 
usual cut-offs. i.e. if prediciton<cut-off, it is a true prediction}
}
\value{
Matthew's Correlation Co-efficient (MCC), Type 1 error, Type 2 error,
raw confusion matrix values and raw values
}
\description{
Efficient function to calculate Matthew's Correlation Co-efficient (MCC)
}
\keyword{internal}
