% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_compute_error.R
\name{compute_error}
\alias{compute_error}
\title{Computing Type 1 Error for Single Cell Expression Case-Control Differential
  Expression Analysis}
\description{
Computes type 1 error for single cell data that is cell-type specifc,
  hierarchical, and compositonal. This function computes type 1 error with the
  single-cell differential expression analysis tool 'MAST', using random
  effects to account for the correlation structure that exists among measures
  from cells within an individual. The type 1 error calculations will borrow
  information from the input data (or the package default data) to simulate
  data under a variety of pre-determined conditions. These conditions include
  foldchange, number of genes, number of samples (i.e., independent
  experimental units), and the mean number of cells per individual.
}
\details{
Prior to running the \code{\link{error_hierarchicell}} function, it
  is important to run the \code{\link{filter_counts}} function followed by
  the \code{\link{compute_data_summaries}} function to build an R object that
  is in the right format for the following simulation function to properly
  work.
}
\note{
Data should be \strong{only for cells of the specific cell-type} you
  are interested in simulating or computing type 1 error for. Data should also
  contain as many unique sample identifiers as possible. If you are inputing
  data that has less than 5 unique values for sample identifier (i.e.,
  independent experimental units), then the empirical estimation of the
  inter-individual heterogeneity is going to be very unstable. Finding such a
  dataset will be difficult at this time, but, over time (as experiments grow
  in sample size and the numbers of publically available single-cell RNAseq
  datasets increase), this should improve dramatically.
}
