% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_compute_power.R
\name{compute_power}
\alias{compute_power}
\title{Computing Power for Single Cell Expression Case-Control Differential
  Expression Analysis}
\description{
Computes power for single cell data that is cell-type specifc,
  hierarchical, and compositonal. This function computes power using random
  effects to account for the correlation structure that exists among measures
  from cells within an individual. The power calculations will borrow
  information from the input data (or the package default data) to simulate
  data under a variety of pre-determined conditions. These conditions include
  foldchange, number of genes, number of samples (i.e., independent
  experimental units), and the mean number of cells per individual.
}
\details{
Prior to running the \code{\link{power_hierarchicell}} function, it
  is important to run the \code{\link{filter_counts}} function followed by
  the \code{\link{compute_data_summaries}} function to build an R object that
  is in the right format for the following simulation function to properly
  work.
}
\note{
Data should be \strong{only for cells of the specific cell-type} you
  are interested in simulating or computing power for. Data should also
  contain as many unique sample identifiers as possible. If you are inputing
  data that has less than 5 unique values for sample identifier (i.e.,
  independent experimental units), then the empirical estimation of the
  inter-individual heterogeneity is going to be very unstable. Finding such a
  dataset will be difficult at this time, but, over time (as experiments grow
  in sample size and the numbers of publically available single-cell RNAseq
  datasets increase), this should improve dramatically.
}
