% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_estimate_parameters.R
\name{empirical_estimation}
\alias{empirical_estimation}
\title{Empirical Estimation of Parameters}
\description{
The most fundamental component of this package is in the
  estimation of the simulation parameters. The functions to estimate
  parameters for the simulation estimate the empirical distributions for
  library size, dropout rate, and global gene means and model the
  hierarchical variance structure of the input data.
}
\details{
Prior to estimating the simulation parameters, it is important to
  run the \code{\link{filter_counts}} function to build a data.frame that is in the right
  format for the following estimation functions to properly compute.
}
\note{
Data should be \strong{only for cells of the specific cell-type} you
  are interested in simulating or computing power for. Data should also
  contain as many unique sample identifiers as possible. If you are inputing
  data that has less than 5 unique values for sample identifier (i.e.,
  independent experimental units), then the empirical estimation of the
  inter-individual heterogeneity is going to be very unstable. Finding such a
  dataset will be difficult at this time, but, over time (as experiments grow
  in sample size and the numbers of publically available single-cell RNAseq
  datasets increase), this should improve dramatically.
}
