% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_estimate_parameters.R
\name{model_dispersion}
\alias{model_dispersion}
\title{Model Dispersion as a function of Intra Means}
\usage{
model_dispersion(data_summaries, plot = FALSE)
}
\arguments{
\item{data_summaries}{an R object that has been output by the package's
compute_data_summaries function.}

\item{plot}{a TRUE/FALSE statement for the output of a plot to observe how
well gene dropout behaves as a function of the intra means}
}
\value{
A vector of length two, where the first number is the shape of the
 gamma distribution and the second number is the rate of the gamma
 distribution
}
\description{
This function estimates the parameters for the gamma distribution
 of dispersion parameters.
}
\details{
Prior to estimating the simulation parameters, it is important to run
 the compute_data_summaries function to build an object that is in the right
 format for the following estimation functions to properly compute.The
 'fitdistrplus' package is required for this function to work. Please install
 it
}
\examples{
clean_expr_data <- filter_counts()
data_summaries <- compute_data_summaries(clean_expr_data)
dispersion_params <- model_dispersion(data_summaries)

}
