% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_estimate_parameters.R
\name{model_drop_sd}
\alias{model_drop_sd}
\title{Model Dropout SD as a Quadratic Function of the Mean Dropout}
\usage{
model_drop_sd(data_summaries, plot = FALSE)
}
\arguments{
\item{data_summaries}{an R object that has been output by the package's
compute_data_summaries function.}

\item{plot}{a TRUE/FALSE statement for the output of a plot to observe how
well gene dropout SD behaves as a quadratic function of mean dropout}
}
\value{
A plot (if plot=TRUE) and a vector of length three, where each
 number is the estimate of the intercept or slope for the model.
}
\description{
This function models gene dropout SD as a quadratic function of the
 dropout mean. Visualizing this fit and looking for oddities and extreme
 outliers may be helpful.
}
\details{
Prior to estimating the simulation parameters, it is important to run
 the compute_data_summaries function to build an object that is in the right
 format for the following estimation functions to properly compute. The
 'ggplot2' package is required for the plotting component of this function to
 work. Please install it.
}
\examples{
\donttest{clean_expr_data <- filter_counts()
data_summaries <- compute_data_summaries(clean_expr_data)
gene_drop_betas <- model_drop_sd(data_summaries)}

}
