% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_estimate_parameters.R
\name{model_gene_drop}
\alias{model_gene_drop}
\title{Model Gene Dropout as a Linear Function of the Grand Mean}
\usage{
model_gene_drop(data_summaries, plot = FALSE)
}
\arguments{
\item{data_summaries}{an R object that has been output by the package's
compute_data_summaries function.}

\item{plot}{a TRUE/FALSE statement for the output of a plot to observe how
well gene dropout behaves as a logit function of the grand mean}
}
\value{
A plot (if plot=TRUE) and a vector of length three, where the first
 number is the estimate of the r value (growth rate) and the second and third
 numbers are the estimate of the initial size and carrying capacity,
 respectively. The estimate of the carrying capacity should be close to one
 and the initial size close to zero.
}
\description{
This function models gene dropout as a logit function of the
 grand mean. Visualizing this fit and looking for oddities and extreme
 outliers may be helpful.
}
\details{
Prior to estimating the simulation parameters, it is important to run
 the compute_data_summaries function to build an object that is in the right
 format for the following estimation functions to properly compute. The
 'ggplot2' package is required for the plotting component of this function to
 work. Please install it.
}
\examples{
data_summaries <- compute_data_summaries(clean_expr_data)
gene_drop_betas <- model_gene_drop(data_summaries)

}
