% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_estimate_parameters.R
\name{model_inter}
\alias{model_inter}
\title{Model Inter-Individual Heterogeneity as a Linear Function of the Grand
 Mean}
\usage{
model_inter(data_summaries, plot = FALSE)
}
\arguments{
\item{data_summaries}{an R object that has been output by the package's
compute_data_summaries function.}

\item{plot}{a TRUE/FALSE statement for the output of a plot to observe how
well inter-individual standard deviation behaves as a linear function of the
grand mean}
}
\value{
A plot (if plot=TRUE) a vector of length two, where the first
 number is the estimate of the intercept for the log model second number is
 the estimate of the slope for the log model.
}
\description{
This function models inter-individual heterogeneity as a log
 function of the grand mean. Visualizing this fit and looking for oddities
 and extreme outliers may be helpful.
}
\details{
Prior to estimating the simulation parameters, it is important to run
 the compute_data_summaries function to build an object that is in the right
 format for the following estimation functions to properly compute. The
 'ggplot2' package is required for the plotting component of this function to
 work. Please install it.
}
\examples{
\donttest{clean_expr_data <- filter_counts()
data_summaries <- compute_data_summaries(clean_expr_data)
inter_betas <- model_inter(data_summaries)}

}
