% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_estimate_parameters.R
\name{model_intra}
\alias{model_intra}
\title{Model Intra-Individual Variance as a Quadratic Function of the Grand
 Mean}
\usage{
model_intra(data_summaries, plot = FALSE)
}
\arguments{
\item{data_summaries}{an R object that has been output by the package's
compute_data_summaries function.}

\item{plot}{a TRUE/FALSE statement for the output of a plot to observe how
well intra-individual standard deviation behaves as a quadratic function of
the grand mean}
}
\value{
A plot (if plot=TRUE) and a vector of length three, where the first
 number is the estimate of beta0 (the intercept) and the second and third
 numbers are the estimates of beta1 and beta2, respectively.
}
\description{
This function models intra-individual standard deviation as a
 quadratic function of the grand mean. Visualizing this fit and looking for
 oddities and extreme outliers may be helpful.
}
\details{
Prior to estimating the simulation parameters, it is important to run
 the compute_data_summaries function to build an object that is in the right
 format for the following estimation functions to properly compute. The
 'ggplot2' package is required for the plotting component of this function to
 work. Please install it.
}
\examples{
data_summaries <- compute_data_summaries(clean_expr_data)
intra_betas <- model_intra(data_summaries)

}
