% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mucosal_cells.R
\name{mucosal_cells}
\alias{mucosal_cells}
\title{Normal Mucosal Cells scRNA-seq data}
\format{A data frame with 160 rows and 57243 variables:
\describe{
  \item{CellID}{Unique cell identifier}
  \item{IndividualID}{Sample identifier}
  \item{Gene}{Generic Gene number}
  ...
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE81nnn/GSE81861}
}
\description{
A dataset containing the gene expression counts for 160 normal mucosal cells
}
